228: Adding security

+++ backend/src/main/java/ch/uzh/ifi/imrg/platform/service/MeetingService.java
@@ -9,6 +9,7 @@ import ch.uzh.ifi.imrg.platform.rest.dto.input.CreateMeetingDTO;
 import ch.uzh.ifi.imrg.platform.rest.dto.output.MeetingOutputDTO;
 import ch.uzh.ifi.imrg.platform.rest.mapper.MeetingsMapper;
 import ch.uzh.ifi.imrg.platform.utils.PatientAppAPIs;
+import ch.uzh.ifi.imrg.platform.utils.SecurityUtil;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import java.util.List;
@@ -27,7 +28,8 @@ public class MeetingService {
 
   private final MeetingRepository meetingRepository;
   private final PatientRepository patientRepository;
-  @PersistenceContext private EntityManager entityManager;
+  @PersistenceContext
+  private EntityManager entityManager;
 
   @Autowired
   public MeetingService(
@@ -37,9 +39,10 @@ public class MeetingService {
     this.patientRepository = patientRepository;
   }
 
-  public Meeting createMeeting(CreateMeetingDTO createMeetingDTO) {
+  public Meeting createMeeting(CreateMeetingDTO createMeetingDTO, String therapistId) {
 
     Patient patient = patientRepository.getPatientById(createMeetingDTO.getPatientId());
+    SecurityUtil.checkOwnership(patient, therapistId);
 
     Meeting meeting = new Meeting();
     meeting.setMeetingStart(createMeetingDTO.getMeetingStart());
@@ -47,12 +50,11 @@ public class MeetingService {
     meeting.setLocation(createMeetingDTO.getLocation());
     meeting.setPatient(patient);
     Meeting createdMeeting = meetingRepository.save(meeting);
-    CreateMeetingDTOPatientAPI createMeetingDTOPatientAPI =
-        new CreateMeetingDTOPatientAPI()
-            .externalMeetingId(createdMeeting.getId())
-            .startAt(meeting.getMeetingStart())
-            .endAt(meeting.getMeetingEnd())
-            .location(meeting.getLocation());
+    CreateMeetingDTOPatientAPI createMeetingDTOPatientAPI = new CreateMeetingDTOPatientAPI()
+        .externalMeetingId(createdMeeting.getId())
+        .startAt(meeting.getMeetingStart())
+        .endAt(meeting.getMeetingEnd())
+        .location(meeting.getLocation());
     PatientAppAPIs.coachMeetingControllerPatientAPI
         .createMeeting1(createMeetingDTO.getPatientId(), createMeetingDTOPatientAPI)
         .block();
@@ -62,24 +64,29 @@ public class MeetingService {
     return createdMeeting;
   }
 
-  public MeetingOutputDTO getMeeting(String meetingId) {
+  public MeetingOutputDTO getMeeting(String meetingId, String therapistId) {
     Meeting meeting = meetingRepository.getReferenceById(meetingId);
+    SecurityUtil.checkOwnership(meeting, therapistId);
+
     return MeetingsMapper.INSTANCE.convertEntityToMeetingOutputDTO(meeting);
   }
 
-  public List<MeetingOutputDTO> getAllMeetingsOfPatient(String patientId) {
+  public List<MeetingOutputDTO> getAllMeetingsOfPatient(String patientId, String therapistId) {
 
     Patient patient = patientRepository.getReferenceById(patientId);
+    SecurityUtil.checkOwnership(patient, therapistId);
 
     return patient.getMeetings().stream()
         .map(MeetingsMapper.INSTANCE::convertEntityToMeetingOutputDTO)
         .collect(Collectors.toList());
   }
 
-  public void deleteMeetingById(String meetingId) {
+  public void deleteMeetingById(String meetingId, String therapistId) {
     Meeting meeting = meetingRepository.getReferenceById(meetingId);
+    SecurityUtil.checkOwnership(meeting, therapistId);
+
     meeting.getPatient().getMeetings().remove(meeting);
     PatientAppAPIs.coachMeetingControllerPatientAPI.deleteMeeting1(
-        meeting.getPatient().getId(), meetingId);
+        meeting.getPatient().getId(), meetingId).block();
   }
 }
